src/mcp-servers/ai-foundry-agent/pyproject.toml
[project]
name = "ai-foundry-mcp-server"
version = "1.0.0"
description = "Model Context Protocol server for Azure AI Foundry Agent with Bing Search"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "azure-ai-projects==1.0.0b3",
    "azure-identity==1.19.0",
    "azure-core==1.31.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
